Moved get_median_count from CountingHash to Hashtable; implemented sweep-reads2.
'sweep-reads2' loads input reads into a hashbits table and then "sweeps" reads
out of another file if they have a median kmer count of 1, i.e. if half or more
of their k-mers are present in the hashbits table.  Since the median k-mer
count function would have been identical in both the Hashbits and CountingHash
subclasses of Hashtable, I promoted it to a base class function.

One use of this is is to use partitions from diginormed data to recover
the full undiginormed members of the partitions from the original data.
It can also be used as an approximate way to pull out reads from one
data set that overlap with reads in another data set, although cumulative
diginorm + partitioning is a better way to do that if you have the memory.